requirements.txt
opencv-python>=4.5.0
deepface>=0.0.79
fer>=22.4.0
mtcnn>=0.1.1
tensorflow>=2.8.0
numpy>=1.19.0
pillow>=9.0.0
matplotlib>=3.4.0
pandas>=1.3.0
tqdm>=4.62.0
google-genai
flask==2.3.3
google-generativeai==0.3.2
mediapipe>=0.10.0
